arkmeds-api/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
factory-boy==3.1.0
Faker==4.14.0
flake8==3.8.4
idna==2.10
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
wrapt==1.12.1
